requirements.txt
fastapi
uvicorn
supabase
pillow
requests
python-multipart
python-dotenv
transformers
torch
numpy
